user,permission edit update apis

+++ app/Http/Controllers/Api/RoleController.php
@@ -14,10 +14,11 @@ class RoleController extends Controller
     //
     function __construct()
     {
-    //      $this->middleware('permission:Role list', ['only' => ['roleList']]);
-    //      $this->middleware('permission:Role create', ['only' => ['roleCreate']]);
-    //      $this->middleware('permission:Role edit', ['only' => ['roleEdit','roleUpdate']]);
-    //      $this->middleware('permission:Role delete', ['only' => ['roleDelete']]);
+        //  $this->middleware('permission:Role list', ['only' => ['roleList']]);
+        //  $this->middleware('permission:Role create', ['only' => ['roleCreate']]);
+        //  $this->middleware('permission:Role edit', ['only' => ['roleEdit','roleUpdate']]);
+        //  $this->middleware('permission:Role delete', ['only' => ['roleDelete']]);
+        //  $this->middleware('permission:Permission assign', ['only' => ['assignPermission']]);
     }
 
     /* 
@@ -123,7 +124,7 @@ class RoleController extends Controller
         }
     
         try {
-            if(Auth::user()->hasPermissionTo('Role list')){
+            if(Auth::user()->hasPermissionTo('Role create')){
                 $role = new Role;
                 $role->name = $request->input('name');
                 $role->guard_name = 'web';
@@ -179,12 +180,20 @@ class RoleController extends Controller
             return response()->json(['status' => false, 'statusCode' => 401, 'message' => $validator->errors()->first()], 401);
         }
         try {
-            $get_role = Role::where('id', $request->role_id)->first();
-            return response()->json([
-                'data' => $get_role,
-                'success' => true,
-                'message' => 'Role detail find successfully'
-            ]);
+            if(Auth::user()->hasPermissionTo('Role edit')){
+                $role = Role::where('id', $request->role_id)->first();
+                return response()->json([
+                    'data' => $role,
+                    'success' => true,
+                    'message' => 'Role detail find successfully'
+                ]);
+            }else{
+                return response()->json([
+                    'data' => [],
+                    'success' => false,
+                    'message' => 'You have not permission to access'
+                ]);
+            }
         } catch (\Throwable $th) {
             return response()->json([
                 'success' => false,
@@ -235,15 +244,24 @@ class RoleController extends Controller
         }
         
         try {
-            $check_role = Role::where('id', $request->role_id)->first();
-            $check_role->name = $request->name;
-            $check_role->guard_name = 'web';
-            $check_role->update();
-            
-            return response()->json([
-                'success' => true,
-                'message' => 'Role update successfully'
-            ]);
+            if(Auth::user()->hasPermissionTo('Role update')){
+                $check_role = Role::where('id', $request->role_id)->first();
+                $check_role->name = $request->name;
+                $check_role->guard_name = 'web';
+                $check_role->update();
+                
+                return response()->json([
+                    'data' => $check_role,
+                    'success' => true,
+                    'message' => 'Role updated successfully'
+                ]);
+            }else{
+                return response()->json([
+                    'data' => [],
+                    'success' => false,
+                    'message' => 'You have not permission to access'
+                ]);
+            }
         } catch (\Throwable $th) {
             return response()->json([
                 'success' => false,
@@ -283,13 +301,22 @@ class RoleController extends Controller
         }
         
         try {
-            $role = Role::where('id', $request->role_id)->first();
-            $role->delete();
-            
-            return response()->json([
-                'success' => true,
-                'message' => 'Role deleted successfully'
-            ]);
+            if(Auth::user()->hasPermissionTo('Role update')){
+                $role = Role::where('id', $request->role_id)->first();
+                $role->delete();
+                
+                return response()->json([
+                    'data' => $role,
+                    'success' => true,
+                    'message' => 'Role deleted successfully'
+                ]);
+            }else{
+                return response()->json([
+                    'data' => [],
+                    'success' => false,
+                    'message' => 'You have not permission to access'
+                ]);
+            }
         } catch (\Throwable $th) {
             return response()->json([
                 'success' => false,
@@ -298,9 +325,52 @@ class RoleController extends Controller
         }
     }
 
+    /*
+    *  @role assign permission
+    *  @param role_id, permission_id
+    *  @response 200 {
+    *   "data": {
+    *    "id": 4,
+    *    "name": "abc",
+    *    "guard_name": "web",
+    *    "created_at": "2023-06-23T06:24:53.000000Z",
+    *    "updated_at": "2023-06-23T06:24:53.000000Z",
+    *    "permissions": [
+    *        {
+    *            "id": 1,
+    *            "name": "Role list",
+    *            "guard_name": "web",
+    *            "created_at": "2023-06-23T06:23:13.000000Z",
+    *            "updated_at": "2023-06-23T06:23:13.000000Z",
+    *            "pivot": {
+    *                "role_id": 4,
+    *                "permission_id": 1
+    *            }
+    *        },
+    *        {
+    *            "id": 3,
+    *            "name": "Role update",
+    *            "guard_name": "web",
+    *            "created_at": "2023-06-23T06:23:13.000000Z",
+    *            "updated_at": "2023-06-23T06:23:13.000000Z",
+    *            "pivot": {
+    *                "role_id": 4,
+    *                "permission_id": 3
+    *            }
+    *        },
+    *    ]
+    *   "success": true,
+    *   "message": "Permission assign successfully",
+    *  }
+    * @response 401 {
+    *   "status": false,
+    *   "statusCode": 401,
+    *   "message": "The name has already been taken."
+    *  }
+    */
+
     public function assignPermission(Request $request)
-    {
-        
+    { 
         $validator = Validator::make($request->all(), [
             'role_id' => 'required|numeric|exists:roles,id',
             'permission_id' => 'required|exists:permissions,id',
@@ -310,15 +380,25 @@ class RoleController extends Controller
         }
 
         try {
-            $role = Role::where('id', $request->role_id)->first();
-            $role->givePermissionTo([
-                $request->permission_id
-            ]);
+            if(Auth::user()->hasPermissionTo('Permission assign')){
+                $role = Role::where('id', $request->role_id)->first();
+                $role->givePermissionTo([
+                    $request->permission_id
+                ]);
+                
+                return response()->json([
+                    'data' => $role,
+                    'success' => true,
+                    'message' => 'Permission assign successfully'
+                ]);
+            }else{
+                return response()->json([
+                    'data' => [],
+                    'success' => false,
+                    'message' => 'You have not permission to access'
+                ]);
+            }
             
-            return response()->json([
-                'success' => true,
-                'message' => 'Permission assign successfully'
-            ]);
         } catch (\Throwable $th) {
             return response()->json([
                 'success' => false,
@@ -326,4 +406,5 @@ class RoleController extends Controller
             ]);
         }
     }
+
 }
